# web/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CUSA SSO API",
    "version": "1.4.0",
    "description": "Custom first-party SSO: Authorization Code + PKCE S256, opaque tokens, application-scoped roles. Not a full OAuth/OIDC provider; no discovery, ID token, refresh token or dynamic registration. Examples are placeholders; never paste production secrets into browser documentation."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same origin as this documentation; production uses HTTPS"
    }
  ],
...
  "paths": {
...
    "/api/sso/token": {
      "post": {
        "operationId": "exchangeCode",
        "summary": "แลก authorization code",
        "description": "เรียกจาก BFF เท่านั้น ต้องมี identity:read และ callback/verifier เดิม ไม่รองรับ refresh_token/password/client_credentials; อย่า retry code เดิมเมื่อไม่ทราบผลการแลก",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "ApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExchangeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                },
                "example": {
                  "access_token": "AAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA",
                  "token_type": "Bearer",
                  "expires_in": 300,
                  "scope": "identity:read profile email"
                }
              }
            }
          },
          "400": {
            "description": "Validation / invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters",
                  "code": "invalid_request"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired API key",
                  "code": "invalid_client"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope / CORS / account access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "API key does not permit this operation",
                  "code": "insufficient_scope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in the permitted scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Resource not found",
                  "code": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Policy version or activity identifier conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Reload the current policy or use the correct event identifier",
                  "code": "CONFLICT"
                }
              }
            }
          },
          "429": {
            "description": "Quota exceeded; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service request quota exceeded",
                  "code": "RATE_LIMITED"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "เวลารอเป็นวินาที ห้าม retry รัว",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service unavailable",
                  "code": "AUDIT_UNAVAILABLE"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unable to process request",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "curl",
            "source": "curl --request POST \"$SSO_ORIGIN/api/sso/token\" \\\n  --header 'Content-Type: application/json' \\\n  --header \"X-API-Key: $SSO_API_KEY\" \\\n  --data '{\"grant_type\":\"authorization_code\",\"code\":\"<CODE>\",\"redirect_uri\":\"https://portal.example.com/auth/callback\",\"code_verifier\":\"<ORIGINAL_VERIFIER>\"}'"
          }
        ]
      }
    },
    "/api/sso/introspect": {
      "post": {
        "operationId": "introspectToken",
        "summary": "ตรวจ token ก่อน protected operation",
        "description": "ต้องมี token:introspect ตรวจได้เฉพาะ token ของ application เดียวกัน HTTP 200 ไม่เท่ากับ authenticated ต้องตรวจ active, aud, exp และ roles; identity cache สูงสุด 5 วินาที",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "ApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntrospectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Active identity หรือ inactive ไม่มีข้อมูลผู้ใช้",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/ActiveToken"
                    },
                    {
                      "$ref": "#/components/schemas/InactiveToken"
                    }
                  ]
                },
                "examples": {
                  "active": {
                    "value": {
                      "sub": "11111111-1111-4111-8111-111111111111",
                      "email": "member@example.com",
                      "name": "Example Member",
                      "given_name": "Example",
                      "family_name": "Member",
                      "department": "ฝ่ายสมาชิก",
                      "roles": [
                        "viewer",
                        "approver"
                      ],
                      "aud": "22222222-2222-4222-8222-222222222222",
                      "active": true,
                      "exp": 2000000000,
                      "scope": "identity:read profile email"
                    }
                  },
                  "inactive": {
                    "value": {
                      "active": false
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation / invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters",
                  "code": "invalid_request"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired API key",
                  "code": "invalid_client"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope / CORS / account access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "API key does not permit this operation",
                  "code": "insufficient_scope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in the permitted scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Resource not found",
                  "code": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Policy version or activity identifier conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Reload the current policy or use the correct event identifier",
                  "code": "CONFLICT"
                }
              }
            }
          },
          "429": {
            "description": "Quota exceeded; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service request quota exceeded",
                  "code": "RATE_LIMITED"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "เวลารอเป็นวินาที ห้าม retry รัว",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service unavailable",
                  "code": "AUDIT_UNAVAILABLE"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unable to process request",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "curl",
            "source": "curl --request POST \"$SSO_ORIGIN/api/sso/introspect\" \\\n  --header 'Content-Type: application/json' \\\n  --header \"X-API-Key: $SSO_API_KEY\" \\\n  --data '{\"token\":\"<ACCESS_TOKEN>\"}'"
          }
        ]
      }
    },
...
    "/api/sso/revoke": {
      "post": {
        "operationId": "revokeServiceSession",
        "summary": "ถอน Token ของเซสชันเฉพาะ Service ของคุณ",
        "description": "BFF ใช้ API key ที่มี token:revoke ส่ง access token ที่ถืออยู่ ระบบถอน access tokens และ authorization codes ของ session เดียวกันเฉพาะ application ของคีย์ ไม่ลบ SSO session ไม่กระทบ Service อื่น ไม่ห้าม login ใหม่ Unknown/cross-app token ตอบ ok เหมือนกัน; cache เดิมอาจมีผลไม่เกิน 5 วินาที ต้องทำลาย BFF session และล้าง cache ของตัวเองด้วย",
        "tags": [
          "Tokens"
        ],
        "security": [
          {
            "ApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IntrospectionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "const": true
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                },
                "example": {
                  "ok": true
                }
              }
            }
          },
          "400": {
            "description": "Validation / invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters",
                  "code": "invalid_request"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired API key",
                  "code": "invalid_client"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope / CORS / account access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "API key does not permit this operation",
                  "code": "insufficient_scope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in the permitted scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Resource not found",
                  "code": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Policy version or activity identifier conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Reload the current policy or use the correct event identifier",
                  "code": "CONFLICT"
                }
              }
            }
          },
          "429": {
            "description": "Quota exceeded; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service request quota exceeded",
                  "code": "RATE_LIMITED"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "เวลารอเป็นวินาที ห้าม retry รัว",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service unavailable",
                  "code": "AUDIT_UNAVAILABLE"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unable to process request",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        },
        "x-codeSamples": [
          {
            "lang": "curl",
            "source": "curl --request POST \"$SSO_ORIGIN/api/sso/revoke\" \\\n  --header 'Content-Type: application/json' \\\n  --header \"X-API-Key: $SSO_API_KEY\" \\\n  --data '{\"token\":\"<ACCESS_TOKEN>\"}'"
          }
        ]
      }
    },
...
    "/api/sso/activity": {
      "post": {
        "operationId": "reportServiceActivity",
        "summary": "Report a real user action from the service backend",
        "description": "Requires a dedicated member:activity API-key scope. Subject must be an active member of the key application. Send only actual user interactions, never polling/introspection/cron. Server time is used; eventId is deduplicated for 30 days. At most one update per user/service per 5 minutes; other new events are acknowledged without being stored. Does not revive a revoked membership. No browser CORS or cookie authentication.",
        "tags": [
          "Identity"
        ],
        "security": [
          {
            "ApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "sub": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "eventId": {
                    "type": "string",
                    "format": "uuid"
                  }
                },
                "required": [
                  "sub",
                  "eventId"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Accepted, duplicate or coalesced activity; see optional flags",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "duplicate": {
                      "type": "boolean"
                    },
                    "recorded": {
                      "const": false
                    },
                    "retryAfter": {
                      "const": 300
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation / invalid grant",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid request parameters",
                  "code": "invalid_request"
                }
              }
            }
          },
          "401": {
            "description": "Missing, expired or revoked credential",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Invalid or expired API key",
                  "code": "invalid_client"
                }
              }
            }
          },
          "403": {
            "description": "Insufficient scope / CORS / account access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "API key does not permit this operation",
                  "code": "insufficient_scope"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found in the permitted scope",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Resource not found",
                  "code": "NOT_FOUND"
                }
              }
            }
          },
          "409": {
            "description": "Policy version or activity identifier conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Reload the current policy or use the correct event identifier",
                  "code": "CONFLICT"
                }
              }
            }
          },
          "429": {
            "description": "Quota exceeded; respect Retry-After",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service request quota exceeded",
                  "code": "RATE_LIMITED"
                }
              }
            },
            "headers": {
              "Retry-After": {
                "description": "เวลารอเป็นวินาที ห้าม retry รัว",
                "schema": {
                  "type": "integer",
                  "minimum": 1
                }
              }
            }
          },
          "503": {
            "description": "Dependency or audit unavailable; fail closed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Service unavailable",
                  "code": "AUDIT_UNAVAILABLE"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "error": "Unable to process request",
                  "code": "INTERNAL_ERROR"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Application-bound backend secret. Scope identity:read, token:introspect, token:revoke or member:activity as required."
      },
      "CookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "__Host-cusa_session",
        "description": "HttpOnly Secure cookie in production; cusa_session on localhost development"
      },
      "CsrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "BearerToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "description": "ข้อความที่ปลอดภัยสำหรับแสดงผล"
          },
          "code": {
            "type": "string",
            "description": "Machine-readable error code; case-sensitive"
          },
          "requestId": {
            "type": "string",
            "description": "Server-generated correlation UUID ถ้ามี",
            "format": "uuid"
          }
        }
      },
...
      "InactiveToken": {
        "type": "object",
        "properties": {
          "active": {
            "type": "boolean",
            "const": false
          }
        },
        "required": [
          "active"
        ],
        "additionalProperties": false
      },
      "ExchangeRequest": {
        "type": "object",
        "properties": {
          "grant_type": {
            "type": "string",
            "enum": [
              "authorization_code"
            ]
          },
          "code": {
            "type": "string",
            "description": "Opaque credential: CSPRNG 32 bytes, base64url without padding",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "redirect_uri": {
            "type": "string",
            "description": "Callback เดียวกับ authorize",
            "minLength": 1,
            "maxLength": 2048
          },
          "code_verifier": {
            "type": "string",
            "description": "PKCE verifier เดิม ห้ามสร้างใหม่ที่ callback",
            "pattern": "^[A-Za-z0-9._~-]{43,128}$"
          }
        },
        "required": [
          "grant_type",
          "code",
          "redirect_uri",
          "code_verifier"
        ],
        "additionalProperties": false
      },
      "TokenResponse": {
        "type": "object",
        "properties": {
          "access_token": {
            "type": "string",
            "description": "Opaque credential: CSPRNG 32 bytes, base64url without padding",
            "pattern": "^[A-Za-z0-9_-]{43}$"
          },
          "token_type": {
            "type": "string",
            "enum": [
              "Bearer"
            ]
          },
          "expires_in": {
            "type": "integer",
            "const": 300,
            "description": "อายุสูงสุด; session ต้นทางอาจหมดอายุก่อน"
          },
          "scope": {
            "type": "string",
            "description": "Space-separated scopes actually approved by the user"
          }
        },
        "required": [
          "access_token",
          "token_type",
          "expires_in",
          "scope"
        ],
        "additionalProperties": false
      },
      "IntrospectionRequest": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string",
            "description": "Opaque access token",
            "minLength": 1,
            "maxLength": 512
          }
        },
        "required": [
          "token"
        ],
        "additionalProperties": false
      },
...
    }
  }
}